requirements.txt
faiss-cpu==1.8.0
numpy==1.26.4
openai==1.34.0
packaging==24.1
pandas==2.2.2
